requirements.txt
absl-py==1.4.0
apache-beam==2.57.0
appnope==0.1.4
array-record==0.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
avro-python3==1.10.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==6.1.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.2
decorator==5.1.1
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
etils==1.3.0
executing==2.0.1
fastavro==1.9.5
fasteners==0.19
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h5py==3.11.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
immutabledict==4.2.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
joblib==1.4.2
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.14
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
lvis==0.5.3
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
# object-detection @ file:./Tensorflow/models/research
objsize==0.7.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.6
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.10.1
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pydot==1.4.2
Pygments==2.18.0
pyjsparser==2.7.1
pymongo==4.8.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.19.0
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io==0.34.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.13.0
termcolor==2.4.0
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.13.2
tf-slim==1.1.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
zstandard==0.23.0
